move getDecimal to util
