+++ packages/xchain-ethereum/CHANGELOG.md
@@ -1,5 +1,11 @@
 # v.x.x.x
 
+# v.0.15.1 (2021-03-23)
+
+### Breaking change
+
+- move `getDecimal` from client to util
+
 # v.0.15.0 (2021-03-23)
 
 ### Add
