+++ packages/xchain-ethereum/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@xchainjs/xchain-ethereum",
-  "version": "0.15.0",
+  "version": "0.15.1",
   "description": "Ethereum client for XChainJS",
   "keywords": [
     "XChain",
